Pipes

+++ Userland/SampleCodeModule/Tests/test_processes.c
@@ -10,9 +10,9 @@ typedef struct P_rq{
 }p_rq;
 
 int64_t test_processes(uint64_t argc, char *argv[]){
-    uint8_t rq;
-    uint8_t alive = 0;
-    uint8_t action;
+    uint64_t rq;
+    uint64_t alive = 0;
+    uint64_t action;
     uint64_t max_processes;
     char * argvAux[] = {"endlessLoop"};
